PN-16170

+++ PAS_Database/dbo/User Defined Types/PostStocklineBatchType.sql
@@ -1,7 +1,7 @@
 ﻿CREATE TYPE [dbo].[PostStocklineBatchType] AS TABLE (
     [StocklineId]     BIGINT          NOT NULL,
-    [Qty]             INT             NOT NULL,
-    [Amount]          DECIMAL (18, 2) NULL,
+    [Qty]             DECIMAL (18, 6) NOT NULL,
+    [Amount]          DECIMAL (18, 6) NULL,
     [ModuleName]      VARCHAR (256)   NULL,
     [UpdateBy]        VARCHAR (256)   NULL,
     [MasterCompanyId] INT             NULL,
